% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{createEdgesDf}
\alias{createEdgesDf}
\title{Create edge dataframe}
\usage{
createEdgesDf(phylo, df.nodes)
}
\arguments{
\item{phylo}{(ape format)}

\item{df.nodes}{(dataframe)}
}
\value{
dataframe
}
\description{
Initialize and fill the edge dataframe of the given phylogeny.
}
