% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{createPhyloBiSSE}
\alias{createPhyloBiSSE}
\title{Create one phylogeny under the BiSSE model}
\usage{
createPhyloBiSSE(n.tips, params)
}
\arguments{
\item{n.tips}{number of tips}

\item{params}{list
\itemize{
\item \verb{$lambda0}: speciation rate of state 0
\item \verb{$q}: transition rate
}}
}
\value{
phylogeny (ape format)
}
\description{
The BiSSE diversification model has 6 parameters:
\itemize{
\item \eqn{\lambda_0} speciation rate of state 0
\item \eqn{\lambda_1} speciation rate of state 1
\item \eqn{\mu_0} extinction rate of state 0
\item \eqn{\mu_1} extinction rate of state 1
\item \eqn{q_{01}} transition rate from state 0 to 1
\item \eqn{q_{10}} transition rate from state 1 to 0
}
}
\details{
We assume 4 constraints:
\itemize{
\item \eqn{\lambda_1 = 2\lambda_0}
\item \eqn{\mu_0 = 0}
\item \eqn{\mu_1 = 0}
\item \eqn{q_{10} = q_{01}}
}
}
\examples{
n.tips <- 10 # 10 extant species
lambda0 <- 1 # speciation rate
q <- 0.1 # transition rate
createPhyloBiSSE(n.tips, list(lambda0 = lambda0, q = q))
}
