% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{drawParamBiSSE}
\alias{drawParamBiSSE}
\title{Draw parameters for the BiSSE model}
\usage{
drawParamBiSSE(params.lim)
}
\arguments{
\item{params.lim}{parameters boundaries (list of vector)
\itemize{
\item \code{lambda0}: boundaries of speciation rate of state 0
\item \code{q}: boundaries of transition rate
}}
}
\value{
list of parameters
}
\description{
Draw speciation rate of state 0 (lambda0) and transition rate randomly.
Other parameters are constrained.
}
