% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{drawParamCRBD}
\alias{drawParamCRBD}
\title{Draw parameters for the CRBD model}
\usage{
drawParamCRBD(lambda.lim)
}
\arguments{
\item{lambda.lim}{speciation rate boundaries (vector)}
}
\value{
list of parameters
}
\description{
To ensure that \eqn{\lambda >> \mu}, parameters are drawn as follow:
\itemize{
\item \eqn{\lambda} in (\code{min(lambda.lim)}, \code{max(lambda.lim)})
\item \eqn{\mu} in (0, \eqn{0.9 \lambda})
}
}
