% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{drawPhyloSize}
\alias{drawPhyloSize}
\title{Draw size of a phylogeny}
\usage{
drawPhyloSize(tips.lim)
}
\arguments{
\item{tips.lim}{size boundaries (vector)}
}
\value{
int
}
\description{
Draw size of a phylogeny
}
