% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{extractLadders}
\alias{extractLadders}
\title{Extract ladders from vector}
\usage{
extractLadders(ladder)
}
\arguments{
\item{ladder}{vector output of \code{\link{writeLadders}}}
}
\value{
list
}
\description{
A node is in a ladder if one of its children is a tip AND the other is not.
Ladder of size one are deleted.
}
