% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesAll}
\alias{fillEdgesAll}
\title{Fill all columns of edge dataframe}
\usage{
fillEdgesAll(df.edges, phylo, df.nodes)
}
\arguments{
\item{df.edges}{dataframe}

\item{phylo}{phylogeny (ape format)}

\item{df.nodes}{dataframe}
}
\value{
dataframe
}
\description{
Fill all columns of edge dataframe
}
