% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesIndex}
\alias{fillEdgesIndex}
\title{Fill the "index" column of edge dataframe}
\usage{
fillEdgesIndex(df.edges, phylo)
}
\arguments{
\item{df.edges}{dataframe created by \code{\link{createEdgesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "index" column of the edge dataframe with their respective index.
}
\seealso{
\code{\link{createEdgesDf}}
}
