% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesIsExt}
\alias{fillEdgesIsExt}
\title{Fill the "isext" column of the edge dataframe}
\usage{
fillEdgesIsExt(df.edges, phylo)
}
\arguments{
\item{df.edges}{dataframe}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
An edge is said externed if it is linked to at least one tip.
The column is filled with logicals.
If the edge is external a \code{TRUE} is inserted, else it is a \code{FALSE}.
}
