% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesLength}
\alias{fillEdgesLength}
\title{Fill the "length" column of edge dataframe}
\usage{
fillEdgesLength(df.edges, phylo)
}
\arguments{
\item{df.edges}{dataframe}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Insert edge lengths in the adequate cells of the edge dataframe.
Edge length corresponds to the time separating the parent from its child.
}
