% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesParentChild}
\alias{fillEdgesParentChild}
\title{Fill the "parent" and "child" columns of edge dataframe}
\usage{
fillEdgesParentChild(df.edges, phylo)
}
\arguments{
\item{df.edges}{dataframe}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
An edge link two species: a parent to its child.
As time goes, the edge goes from parent to child.
}
