% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillEdgesPart}
\alias{fillEdgesPart}
\title{Fill the "part" column of edge dataframe}
\usage{
fillEdgesPart(df.edges, df.nodes)
}
\arguments{
\item{df.edges}{dataframe}

\item{df.nodes}{dataframe}
}
\value{
dataframe
}
\description{
The phylogeny is sliced into 3 equal parts.
The part of the edge is determined by the part of the child,
e.g. if the edge goes from a parent belonging to part 3 and to a child
belonging to part 2, then the edge belongs to part 2.
}
