% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesAll}
\alias{fillNodesAll}
\title{Fill all columns of node dataframe}
\usage{
fillNodesAll(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill all columns of node dataframe
}
