% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesColless}
\alias{fillNodesColless}
\title{Fill the "colless" column of node dataframe}
\usage{
fillNodesColless(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "colless" column of the nodes dataframe with corresponding colless
score. The colless score is equal to the absolute difference of the number
of tips on the left and right side of the node.
For more detail, see
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}.
}
\seealso{
\code{\link{createNodesDf}},
\code{\link{getNodeColless}}
}
