% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesDepth}
\alias{fillNodesDepth}
\title{Fill the "depth" column of node dataframe}
\usage{
fillNodesDepth(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "depth" column of the nodes dataframe with corresponding
depth. The depth is the number of edges between the node and the root.
For instance, for the following simple phylogeny root--->--->B,
depth(A) = 1 and depth(B) = 2
For more detail, see
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}.
}
\seealso{
\code{\link{createNodesDf}},
}
