% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesDist}
\alias{fillNodesDist}
\title{Fill the "dist" column of node dataframe}
\usage{
fillNodesDist(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "dist" column of the nodes dataframe with corresponding distances
to root.
}
\seealso{
\code{\link{createNodesDf}},
\code{\link[castor]{get_all_distances_to_root}}
}
