% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesIndex}
\alias{fillNodesIndex}
\title{Fill the "index" column of node dataframe}
\usage{
fillNodesIndex(df.nodes)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}
}
\value{
dataframe
}
\description{
Fill the "index" column of the nodes dataframe with their respective index.
}
\seealso{
\code{\link{createNodesDf}}
}
