% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesIsTip}
\alias{fillNodesIsTip}
\title{Fill the "istip" column of node dataframe}
\usage{
fillNodesIsTip(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "istip" column of the nodes dataframe with corresponding logicals.
If the node is a tip \code{TRUE}, else \code{FALSE}.
}
\seealso{
\code{\link{createNodesDf}}
}
