% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesPart}
\alias{fillNodesPart}
\title{Fill the "part" column of node dataframe}
\usage{
fillNodesPart(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "part" column of the nodes dataframe with corresponding phylogeny
part. The phylogeny is cut in 3 equals regarding time.
Let's be D>0 the phylogeny height (maximal distance to the root),
then nodes with distance to the root d such that:
\itemize{
\item d<D/3 belongs to part 1;
\item D/3<d<2D/3 belongs to part 2;
\item d>2D/3 belongs to part 3.
For more detail, see
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}.
}
}
\seealso{
\code{\link{createNodesDf}},
}
