% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{fillNodesStair}
\alias{fillNodesStair}
\title{Fill the "stair" column of node dataframe}
\usage{
fillNodesStair(df.nodes, phylo)
}
\arguments{
\item{df.nodes}{dataframe created by \code{\link{createNodesDf}}}

\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Fill the "stair" column of the nodes dataframe with corresponding
stairecaseness. The stairecaseness is equal to the ratio of the number
of tips on each side.
For more detail, see
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}.
}
\seealso{
\code{\link{createNodesDf}},
\code{\link{getNodeStair}}
}
