% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getBranchLengthStats}
\alias{getBranchLengthStats}
\title{Compute branch lengths statistics}
\usage{
getBranchLengthStats(branch_length)
}
\arguments{
\item{branch_length}{vector}
}
\value{
list
\itemize{
\item \verb{$mean}: branch lengths mean
\item \verb{$median}: branch lengths median
\item \verb{$var}: branch length variance
}
}
\description{
Compute the mean, median and variance of a list of branch lengths.
}
