% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getChildLeftRight}
\alias{getChildLeftRight}
\title{Left and right child of a node}
\usage{
getChildLeftRight(phylo, i)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}
}
\value{
list
\itemize{
\item \verb{$left}: node left child
\item \verb{$right}: node right child
}
}
\description{
The left children is the node the further from the root, the right children
is the closest to the root.
If node is a tip returns NA.
}
