% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getImbalancedNodesProp}
\alias{getImbalancedNodesProp}
\title{Proportion of imbalanced nodes}
\usage{
getImbalancedNodesProp(df.nodes)
}
\arguments{
\item{df.nodes}{dataframe}
}
\value{
numeric
}
\description{
A node is imbalanced if it has a different number of left and right tips.
Equivalently, the node has a stairecaseness different from 1.
}
