% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getInLadderNodesProp}
\alias{getInLadderNodesProp}
\title{Proportion of internal nodes in ladder}
\usage{
getInLadderNodesProp(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
numeric
}
\description{
A node is in a ladder if one of its children is a tip AND the other is not.
Ladder should be at least of size two.
}
\seealso{
\code{\link{writeLadders}}, \code{\link{extractLadders}},
\code{\link{getLadders}}
}
