% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getLadders}
\alias{getLadders}
\title{Get ladders of a phylogeny}
\usage{
getLadders(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
list of ladders
}
\description{
A node is in a ladder if one of its children is a tip AND the other is not.
Ladder should be at least of size two.
}
\seealso{
\code{\link{writeLadders}}, \code{\link{extractLadders}}
}
