% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getLttCoords}
\alias{getLttCoords}
\title{Sample 20 LTT coordinates}
\usage{
getLttCoords(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
list
\itemize{
\item \verb{$LTT_coordX} with \code{coord} either \code{t} for time or \code{N} for number of
lineages and \code{X} between 1 and 20
\item for instance, \verb{$LTT_N15} is the number of lineages for bin 15 out of 20
}
}
\description{
Take 20 points uniformly distributed regarding the number of lineages.
Save LTT coordinates for these 20 points.
Phylogeny needs to be at least to have 20 tips.
}
