% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getLttSlopes}
\alias{getLttSlopes}
\title{Compute 10 LTT slopes}
\usage{
getLttSlopes(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
list
}
\description{
Divide the tree in 3 equals part (trough time) and compute the slope for each
part (named slope$p.i). Then do the ratio of:
\itemize{
\item slope of the 2nd over the slope of the 1st part
\item slope of the 3rd over the slope of the 2nd part
}
}
