% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getMaxWidthDiff}
\alias{getMaxWidthDiff}
\title{Maximal consecutive width difference of a phylogeny}
\usage{
getMaxWidthDiff(df.nodes)
}
\arguments{
\item{df.nodes}{data.frame}
}
\value{
numeric
}
\description{
Width at a given depth is the number of nodes at the considered depth.
The width of the phylogeny is the maximum of the widths for all possible
depths.
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
