% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getNodeColless}
\alias{getNodeColless}
\title{Compute the colless score of a node}
\usage{
getNodeColless(phylo, i)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}
}
\value{
int
}
\description{
The colless score is defined as the difference of the number of left and
right tips of a node. Thus, this score can only be computed for internal
nodes (i.e. with at least one descendant).
For more detail, see
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}.
}
