% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getNodeStair}
\alias{getNodeStair}
\title{Compute the stairecaseness of a node}
\usage{
getNodeStair(phylo, i)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}
}
\value{
float
}
\description{
Compute the ratio between the minimum and maximum number of tips on each side
for a node. Can be computed only for internal node.
}
