% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{getSackin}
\alias{getSackin}
\title{Sackin score}
\usage{
getSackin(df.nodes)
}
\arguments{
\item{df.nodes}{data.frame}
}
\value{
int
}
\description{
Sum of the depth over the phylogeny tips.
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
