% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{initializeEdgesDf}
\alias{initializeEdgesDf}
\title{Create edge dataframe}
\usage{
initializeEdgesDf(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Create a dataframe to fill with edge attributes. Rows correspond to edges,
column to attributes. All cells are filled with NAs.
}
