% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{initializeNodesDf}
\alias{initializeNodesDf}
\title{Create node dataframe}
\usage{
initializeNodesDf(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
dataframe
}
\description{
Create a dataframe to fill with node attributes. Rows correspond to nodes,
column to attributes. All cells are filled with NAs.
}
