% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{isInLadder}
\alias{isInLadder}
\title{Is the node in a ladder?}
\usage{
isInLadder(phylo, i)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}
}
\value{
logical
}
\description{
A node is in a ladder if one of its children is a tip AND the other is not.
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
