% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isTip}
\alias{isTip}
\title{Is a phylogeny node a tip?}
\usage{
isTip(phylo, i)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}
}
\value{
logical
}
\description{
Is a phylogeny node a tip?
}
