% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numberEdges}
\alias{numberEdges}
\title{Compute number of edges of a phylogeny}
\usage{
numberEdges(phylo)
}
\arguments{
\item{phylo}{(ape format)}
}
\value{
int
}
\description{
Compute number of edges of a phylogeny
}
