% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numberNodesInternal}
\alias{numberNodesInternal}
\title{Compute number of internal nodes of a phylogeny}
\usage{
numberNodesInternal(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
int
}
\description{
Compute number of internal nodes of a phylogeny
}
