% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numberNodesTotal}
\alias{numberNodesTotal}
\title{Compute total number of nodes of a phylogeny}
\usage{
numberNodesTotal(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
int
}
\description{
Compute total number of nodes of a phylogeny
}
