% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numberTips}
\alias{numberTips}
\title{Compute number of tips of a phylogeny}
\usage{
numberTips(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
int
}
\description{
Compute number of tips of a phylogeny
}
