% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyloBiSSE}
\alias{phyloBiSSE}
\title{Attribute of 10 phylogenies created with the BiSSE model}
\format{
A list of phylogeny data create with the BiSSE model
\describe{
\item{\verb{$phylo}}{phylogenies (ape format)}
\item{\verb{$params}}{phylogeny parameters (dataframe)}
\item{\verb{$sumstat}}{phylogeny summary statistics (dataframe)}
...
}
}
\source{
see \verb{/data-raw/phyloBiSSE.R}
}
\usage{
phyloBiSSE
}
\description{
A dataset containing 10 phylogenies attributes
}
\keyword{datasets}
