% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phyloCRBD}
\alias{phyloCRBD}
\title{Attribute of 10 phylogenies created with the CRBD model}
\format{
A list of phylogeny data create with the CRBD model
\describe{
\item{\verb{$phylo}}{phylogenies (ape format)}
\item{\verb{$params}}{phylogeny parameters (dataframe)}
\item{\verb{$sumstat}}{phylogeny summary statistics (dataframe)}
...
}
}
\source{
see \verb{/data-raw/phyloCRBD.R}
}
\usage{
phyloCRBD
}
\description{
A dataset containing 10 phylogenies attributes
}
\keyword{datasets}
