% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{simulatePhylo}
\alias{simulatePhylo}
\title{Backbone to simulate phylogenies}
\usage{
simulatePhylo(
  createPhylo,
  drawParam,
  n.phylo,
  tips.lim,
  params.lim,
  sumstat_check,
  verbose
)
}
\arguments{
\item{createPhylo}{function used to create one phylogeny under a diversification model}

\item{drawParam}{function to draw parameters of the diversification model}

\item{n.phylo}{number of phylogenies to generate (int)}

\item{tips.lim}{phylogeny size boundaries}

\item{params.lim}{parameter boundaries}

\item{sumstat_check}{logical}

\item{verbose}{logical}
}
\description{
Internal function used to simulated phylogenies. Called by \code{\link{simulatePhyloCRBD}}
and \code{\link{simulatePhyloBiSSE}}.
}
