% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{simulatePhyloBiSSE}
\alias{simulatePhyloBiSSE}
\title{Simulate phylogenies under the BiSSE model}
\usage{
simulatePhyloBiSSE(
  n.phylo,
  tips.lim,
  params.lim,
  sumstat_check = T,
  verbose = F
)
}
\arguments{
\item{n.phylo}{number of phylogenies to simulate}

\item{tips.lim}{phylogeny size range}

\item{params.lim}{parameter ranges}

\item{sumstat_check}{whether to check that summary statistics can be computed
for each phylogeny and save them (logical)}

\item{verbose}{whether to print progress (logical)}
}
\value{
list
\itemize{
\item \verb{$phylo}: list of simulated phylogenies (ape format)
\item \verb{$params}: dataframe of phylogeny parameters
\item \verb{$sumstat}: dataframe of phylogeny summary statistics
(empty if \code{ss_check = FALSE})
}
}
\description{
Generate a list of phylogenies under the Binary State Speciation and Extinction
(BiSSE) model. Can check and save summary statistics of each phylogeny by
activating \code{sumstat_check}. Moreover true parameters of each phylogeny are also
returned.
}
\examples{
simulatePhyloBiSSE(10, c(50, 100), list(lambda0 = c(0.1, 1.0), q = c(0, 0.1)))
}
