% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{simulatePhyloCRBD}
\alias{simulatePhyloCRBD}
\title{Simulate phylogenies under the CRBD model}
\usage{
simulatePhyloCRBD(
  n.phylo,
  tips.lim,
  lambda.lim,
  sumstat_check = T,
  verbose = F
)
}
\arguments{
\item{n.phylo}{number of phylogenies to simulate}

\item{tips.lim}{phylogeny size range}

\item{lambda.lim}{speciation rate range}

\item{sumstat_check}{whether to check that summary statistics can be computed
for each phylogeny and save them (logical)}

\item{verbose}{whether to print progress (logical)}
}
\value{
list
\itemize{
\item \verb{$phylo}: list of simulated phylogenies (ape format)
\item \verb{$params}: dataframe of phylogeny parameters
\item \verb{$sumstat}: dataframe of phylogeny summary statistics
(empty if \code{ss_check = FALSE})
}
}
\description{
Generate a list of phylogenies under the Constant Rate Birth Death (CRBD)
model. Can check and save summary statistics of each phylogeny by activating
\code{sumstat_check}. Moreover true parameters of each phylogeny are also
returned.
}
\examples{
simulatePhyloCRBD(10, c(50, 100), c(0.1, 1.)) # simulate 10 phylo
}
