% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{splitLttCoords}
\alias{splitLttCoords}
\title{Split the LTT coordinates regarding the number of events}
\usage{
splitLttCoords(coords, n = 10)
}
\arguments{
\item{coords}{ltt coordinates (dataframe)}

\item{n}{number of splits (int, default n=10)}
}
\value{
list
}
\description{
Split the LTT coordinates regarding the number of events
}
