% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{sumStatsBranchLength}
\alias{sumStatsBranchLength}
\title{Compute branch lengths summary statistics}
\usage{
sumStatsBranchLength(df.edges, phylo)
}
\arguments{
\item{df.edges}{dataframe}

\item{phylo}{phylogeny (ape format)}
}
\value{
list
\itemize{
\item \verb{$stat.all}: statistic computed on all branch lengths
(stat is either mean, median or var)
\item \verb{$stat.ext}: statistic computed on external branch lengths only
\item \verb{$stat.inti}: statistic computed on internal branch lengths only
belonging to part i (i is either 1,2 or 3)
\item \verb{$stat.intext1}: \verb{$stat.inti} / \verb{$stat.ext}
}
}
\description{
Compute the 25 summary statistics related to branch lengths listed below.
A branch is external if linked to a tip, other the branch is internal.
The branches are sorted in 3 parts regarding the time of the child.
\tabular{ll}{
Variable \tab Description\cr
height \tab phylogeny height\cr
mean.all \tab mean of all branch lengths\cr
median.all \tab median of all branch lengths\cr
var.all \tab variance of all branch lengths\cr
mean.ext \tab mean of external branch lengths\cr
median.ext \tab median of external branch lengths\cr
var.ext \tab variance of external branch lengths\cr
mean.int1 \tab mean of internal branch lengths of part 1\cr
median.int1 \tab median of internal branch lengths of part 1\cr
var.int1 \tab variance of internal branch lengths of part 1\cr
mean.int2 \tab mean of internal branch lengths of part 2\cr
median.int2 \tab median of internal branch lengths of part 2\cr
var.int2 \tab variance of internal branch lengths of part 2\cr
mean.int3 \tab mean of internal branch lengths of part 3\cr
median.int3 \tab median of internal branch lengths of part 3\cr
var.int3 \tab variance of internal branch lengths of part 3\cr
mean.intext1 \tab mean.int1 / mean.ext\cr
median.int1 \tab median.int1 / median.ext\cr
var.int1 \tab var.int1 / var.ext\cr
mean.intext2 \tab mean.int2 / mean.ext\cr
median.int2 \tab median.int2 / median.ext\cr
var.int2 \tab var.int2 / var.ext\cr
mean.intext3 \tab mean.int3 / mean.ext\cr
median.int3 \tab median.int3 / median.ext\cr
var.int3 \tab var.int3 / var.ext
}
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
