% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{sumStatsLtt}
\alias{sumStatsLtt}
\title{LTT summary statistics}
\usage{
sumStatsLtt(phylo)
}
\arguments{
\item{phylo}{phylogeny (ape format)}
}
\value{
list
the variable listed above can be accessed with \code{$} + variable name.
}
\description{
There are 51 LTT summary statistics listed below.
\tabular{ll}{
Variable \tab Description\cr
LTT_slope\code{i} \tab slope of LTT i-th part in semilog scale (i in [1,10])\cr
LTT_t\code{i} \tab i-th time coordinate of binned LTT (i in [1,20])\cr
LTT_N\code{i} \tab i-th lineage coordinate of binned LTT (i in [1,20])\cr
n_tips \tab phylogeny number of tips
}
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
