% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{sumStatsTopo}
\alias{sumStatsTopo}
\title{Topological summary statistics}
\usage{
sumStatsTopo(phylo, df.nodes)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{df.nodes}{dataframe}
}
\value{
list
the variable listed above can be accessed with \code{$} + variable name.
}
\description{
There are 8 topological summary statistics listed below.
\tabular{ll}{
Variable \tab Description\cr
colless \tab absolute difference in left and right tips, summed over int.
nodes\cr
sackin \tab depth summed over tips\cr
widthdepth \tab ratio of maximal width over maximal depth\cr
deltaw \tab maximal consecutive width difference\cr
maxladder \tab maximal ladder size\cr
inladder \tab proportion of nodes in ladder\cr
imbalance \tab proportion of imbalanced nodes (i.e. colless !=0)\cr
stair \tab ratio of tips on each side summed over internal nodes}
}
\seealso{
\href{https://doi.org/10.1371/journal.pcbi.1005416}{Saulnier et al., 2014, PLOS}
}
