% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{sumStatTopo}
\alias{sumStatTopo}
\title{Topological summary statistics}
\usage{
sumStatTopo(phylo, df.nodes)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{df.nodes}{dataframe}
}
\value{
list
}
\description{
There are 8 topological summary statistics listed below.
}
