% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{writeLadders}
\alias{writeLadders}
\title{Traverse a phylogeny and write ladders}
\usage{
writeLadders(phylo, i, preorder)
}
\arguments{
\item{phylo}{phylogeny (ape format)}

\item{i}{node index}

\item{preorder}{vector}
}
\value{
vector with ladders
}
\description{
Traverse the phylogeny in preoder and write node indexes belonging to
ladders. Nodes that don't belong to ladder are marked with \code{-1}.
A node is in a ladder if one of its children is a tip AND the other is not.
}
\seealso{
\code{\link{isInLadder}}, \code{\link{getChildLeftRight}}
}
