% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryStatistics.R
\name{createNodesDf}
\alias{createNodesDf}
\title{Create node dataframe}
\usage{
createNodesDf(phylo)
}
\arguments{
\item{phylo}{(ape format)}
}
\value{
dataframe
}
\description{
Initialize and fill the node dataframe of the given phylogeny.
}
