% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{createPhyloCRBD}
\alias{createPhyloCRBD}
\title{Create a phylogeny under the CRBD (Constant Rate Birth Death) model}
\usage{
createPhyloCRBD(n.tips, lambda, mu)
}
\arguments{
\item{n.tips}{number of tips}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}
}
\value{
phylogeny (ape format)
}
\description{
Create a phylogeny under the CRBD (Constant Rate Birth Death) model
}
