% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulatePhylogenies.R
\name{createPhyloCRBD}
\alias{createPhyloCRBD}
\title{Create one phylogeny under the Constant Rate Birth-Death (CRBD) model}
\usage{
createPhyloCRBD(n.tips, params)
}
\arguments{
\item{n.tips}{number of tips}

\item{lambda}{speciation rate}

\item{mu}{extinction rate}
}
\value{
phylogeny (ape format)
}
\description{
Create one phylogeny under the Constant Rate Birth-Death (CRBD) model
}
\examples{
n.tips <- 10 # 10 extant species
lambda <- 1 # speciation rate
mu <- 0 # extinction rate
createPhyloCRBD(n.tips, lambda, mu)
}
